requirements.txt
# Python数据抓取工具依赖包
# 网页抓取相关
selenium==4.15.2
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3

# 数据库连接
pymongo==4.6.0

# 定时任务
schedule==1.2.0
apscheduler==3.10.4

# 日志记录
loguru==0.7.2

# 数据处理
pandas==2.1.4
numpy==1.24.3

# 配置文件处理
python-dotenv==1.0.0
pyyaml==6.0.1

# HTTP客户端
httpx==0.25.2

# 时间处理
python-dateutil==2.8.2

# 系统工具
psutil==5.9.6

# WebDriver管理
webdriver-manager==4.0.1